requirements.txt
flask
flask-socketio
selenium
webdriver-manager
psycopg2-binary
